Add require_relative module to class
